adding get profile

+++ Server/controllers/profileController.js
@@ -1,7 +1,5 @@
 const Profile = require('../models/profileModel');
-
-
-
+const fs=require('fs');
 const validateFields = ({ firstName, lastName, filiere, experiences, posteActuel, experiencesPassee, niveau,role,promotion }) => {
   if (!firstName || !lastName || !filiere) {
     throw new Error('Tous les champs obligatoires doivent être remplis.');
@@ -49,28 +47,35 @@ exports.updateProfile = async (req, res) => {
     if (!userId) {
       return res.status(400).json({ message: 'userId and image are required in the request.' });
     }
-    let new_image='';
-    if(req.file){
-      new_image=req.file?req.file.filename:0;
-      try{
-        fs.unlinkSync('./uploads/'+req.body.old_image);
-      }catch(err){
-        console.log(err);
+
+    let new_image = '';
+    if (req.file) {
+      new_image = req.file ? req.file.filename : '';
+      if (req.body.old_image) {
+        try {
+          fs.unlinkSync('./uploads/' + req.body.old_image);
+        } catch (err) {
+          console.log(err);
+        }
       }
-   }else{
-    new_image=req.body.old_image;
-   }
-    const { firstName, lastName, filiere, niveau, experiences, posteActuel, experiencesPassee} = req.body;
-    validateFields({ firstName, lastName, filiere, niveau, experiences, posteActuel, experiencesPassee});
+    } else {
+      new_image = req.body.old_image;
+    }
+
+    const { firstName, lastName, filiere, niveau, experiences, posteActuel, experiencesPassee } = req.body;
+    validateFields({ firstName, lastName, filiere, niveau, experiences, posteActuel, experiencesPassee });
+
     const existingProfile = await Profile.findOne({ userId });
 
     if (!existingProfile) {
       return res.status(404).json({ message: 'Profile not found for the given userId.' });
     }
+
     existingProfile.firstName = firstName;
     existingProfile.lastName = lastName;
     existingProfile.filiere = filiere;
-    existingProfile.image= new_image;
+    existingProfile.image = new_image;
+
     if (role === 'etudiant') {
       existingProfile.experiences = experiences;
       existingProfile.posteActuel = undefined;
@@ -79,7 +84,7 @@ exports.updateProfile = async (req, res) => {
     } else if (role === 'laureat') {
       existingProfile.posteActuel = posteActuel;
       existingProfile.experiencesPassee = experiencesPassee;
-      newProfile.promotion=promotion;
+      existingProfile.promotion = promotion; 
       existingProfile.experiences = undefined;
       existingProfile.niveau = undefined; 
     }
@@ -92,6 +97,7 @@ exports.updateProfile = async (req, res) => {
   }
 };
 
+
 exports.checkProfile = async (req, res) => {
   try {
     const userId = req.params.userId; 
@@ -100,6 +106,20 @@ exports.checkProfile = async (req, res) => {
       res.json({ hasProfile: true });
     } else {
       res.json({ hasProfile: false });
+    }
+  } catch (error) {
+    console.error('Error checking profile:', error);
+    res.status(500).json({ error: 'Internal Server Error' });
+  }
+};
+exports.getProfile = async (req, res) => {
+  try {
+    const userId = req.params.userId; 
+    const profile = await Profile.findOne({ userId });
+    if (profile) {
+      res.json({profile });
+    } else {
+      res.json({message:"error in getting profile"})
     }
   } catch (error) {
     console.error('Error checking profile:', error);
